+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### 4.0.1 (unreleased)
+* fix external storage write for API 29
+* combine MAIN and DEBUG views
+
 ### 4.0.0 (27 Aug 2020)
 * Reconfigure app to be Scanner only, no jammer components.
 * Preliminary alpha type release.